// CoreUI/src/app/setup/models/printer.setup.model.ts
export class PrinterSetup {
    PrinterIPAddress: string;
    ReceiptLogoPath: string;
    ReceiptAddress: string;
    ReceiptEmail: string;
    ReceiptPhone: string;
    NTN: string;
    TaxName: string;
    FooterNotes: string;
    WebSite: string;
    ImageFile:string;
    branchName: string; 
}